requirements.txt
scikit-learn
requests
python-dotenv
schedule
pandas

#install dengan perintah pip install -r requirements.txt